+++ R/upperCamelCase-methods.R
@@ -12,15 +12,6 @@ NULL
 
 
 
-#' @rdname upperCamelCase
-#' @name upperCamelCase
-#' @importFrom syntactic upperCamelCase
-#' @usage upperCamelCase(object, ...)
-#' @export
-NULL
-
-
-
 `upperCamelCase,atomic` <-  # nolint
     function(object, names = TRUE, strict = FALSE) {
         assert(